# backend/almacenamiento/entrega/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CloudVault — contrato de almacenamiento de Dani",
    "version": "fase-09",
    "description": "Contrato ejecutable basado en Contratos de API - CloudVault.pdf. Inicio, confirmación y descarga instalados y probados con servicios sintéticos; mantenimiento interno y periódico propio disponible. Proveedor real, SQL compartido y ejecución periódica en el ambiente compartido pendientes."
  },
...
  "paths": {
    "/api/v1/archivos/iniciar-carga/": {
      "post": {
        "operationId": "iniciar_carga",
        "tags": [
          "Almacenamiento"
        ],
        "x-estado-implementacion": "endpoint-implementado-integracion-pendiente",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Respuesta definida por el PDF vigente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IniciarCargaSuccess"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "NO_AUTENTICADO, TOKEN_INVALIDO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              },
              "WWW-Authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "SIN_PERMISO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "NO_ENCONTRADO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "405": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "CUOTA_EXCEDIDA",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              },
              "Retry-After": {
                "description": "Segundos hasta reintentar cuando el limitador los conoce.",
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              }
            }
          },
          "500": {
            "description": "ERROR_INTERNO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "SERVICE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IniciarCargaInput"
              }
            }
          }
        }
      }
    },
    "/api/v1/archivos/{id}/confirmar-carga/": {
      "post": {
        "operationId": "confirmar_carga",
        "tags": [
          "Almacenamiento"
        ],
        "x-estado-implementacion": "endpoint-implementado-integracion-pendiente",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta definida por el PDF vigente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmarCargaSuccess"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "NO_AUTENTICADO, TOKEN_INVALIDO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              },
              "WWW-Authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "SIN_PERMISO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "NO_ENCONTRADO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "405": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "CUOTA_EXCEDIDA",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "ERROR_INTERNO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "SERVICE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmarCargaInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
            }
          }
        ]
      }
    },
    "/api/v1/archivos/{id}/descarga/": {
      "get": {
        "operationId": "descarga",
        "tags": [
          "Almacenamiento"
        ],
        "x-estado-implementacion": "endpoint-implementado-integracion-pendiente",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta definida por el PDF vigente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DescargaSuccess"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "NO_AUTENTICADO, TOKEN_INVALIDO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              },
              "WWW-Authenticate": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "SIN_PERMISO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "NO_ENCONTRADO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "405": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              },
              "Retry-After": {
                "description": "Segundos hasta reintentar cuando el limitador los conoce.",
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              }
            }
          },
          "500": {
            "description": "ERROR_INTERNO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "SERVICE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-referrer"
                  ]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
            }
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "IniciarCargaInput": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "pattern": "^(?!\\s*$)(?!\\s*\\.{1,2}\\s*$)[^/\\\\\\x00-\\x1f\\x7f]+$"
          },
          "tamano_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 1073741824,
            "description": "Límite operativo del perfil documentado; la cuota se valida antes de firmar."
          },
          "tipo_mime": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[A-Za-z0-9!#$&^_.+\\-]+/[A-Za-z0-9!#$&^_.+\\-]+(?![\\s\\S])"
          },
          "carpeta_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
            "nullable": true,
            "default": null
          }
        },
        "additionalProperties": false,
        "required": [
          "nombre",
          "tamano_bytes",
          "tipo_mime"
        ]
      },
      "ConfirmarCargaInput": {
        "type": "object",
        "properties": {
          "etag": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "pattern": "^[^\\x00-\\x1f\\x7f]+$",
            "description": "Opcional; se conserva literalmente y no equivale a SHA-256."
          }
        },
        "additionalProperties": false
      },
      "IniciarCargaSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "archivo_id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
              },
              "url_subida": {
                "type": "string",
                "format": "uri",
                "maxLength": 8192
              },
              "metodo": {
                "type": "string",
                "enum": [
                  "PUT"
                ]
              },
              "encabezados": {
                "type": "object",
                "properties": {
                  "Content-Type": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "pattern": "^[A-Za-z0-9!#$&^_.+\\-]+/[A-Za-z0-9!#$&^_.+\\-]+(?![\\s\\S])"
                  }
                },
                "additionalProperties": false,
                "required": [
                  "Content-Type"
                ]
              },
              "expira_en": {
                "type": "string",
                "format": "date-time",
                "description": "ISO 8601 en UTC."
              }
            },
            "additionalProperties": false,
            "required": [
              "archivo_id",
              "url_subida",
              "metodo",
              "encabezados",
              "expira_en"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "data"
        ]
      },
      "ConfirmarCargaSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
              },
              "nombre": {
                "type": "string",
                "minLength": 1,
                "maxLength": 255,
                "pattern": "^(?!\\s*$)(?!\\s*\\.{1,2}\\s*$)[^/\\\\\\x00-\\x1f\\x7f]+$"
              },
              "es_nuevo": {
                "type": "boolean"
              },
              "en_papelera": {
                "type": "boolean"
              }
            },
            "additionalProperties": false,
            "required": [
              "id",
              "nombre",
              "es_nuevo",
              "en_papelera"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "data"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "NO_AUTENTICADO",
                  "TOKEN_INVALIDO",
                  "SIN_PERMISO",
                  "NO_ENCONTRADO",
                  "CUOTA_EXCEDIDA",
                  "RATE_LIMITED",
                  "ERROR_INTERNO",
                  "SERVICE_UNAVAILABLE"
                ]
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "additionalProperties": false,
            "required": [
              "code",
              "fields"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}